requirements.txt
discord.py
yt-dlp
python-dotenv
spotipy
